# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Plateforme d'Évaluation de la Maturité des DSIN
  version: 1.0.0
  description: Documentation API pour l'application d'évaluation de la maturité DevSecOps
  contact:
    name: Support API
    email: support@example.com
servers:
  - url: http://localhost:5000/api
    description: Serveur de développement
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Application:
      type: object
      properties:
        id_application:
          type: string
          format: uuid
          description: Identifiant unique de l'application
        nom_application:
          type: string
          maxLength: 100
          description: Nom de l'application
        statut:
          type: string
          enum:
            - Projet
            - Run
          description: Statut de l'application
        type:
          type: string
          enum:
            - Build
            - Buy
          description: Type d'application (développée en interne ou achetée)
        hebergement:
          type: string
          enum:
            - Cloud
            - Prem
            - Hybrid
          description: Type d'hébergement de l'application
        architecture_logicielle:
          type: string
          enum:
            - ERP
            - Multitenant SAAS
            - MVC
            - Monolithique
          description: Architecture logicielle de l'application
        date_creation:
          type: string
          format: date-time
          description: Date de création de l'enregistrement
        date_modification:
          type: string
          format: date-time
          description: Date de dernière modification de l'enregistrement
        date_mise_en_prod:
          type: string
          format: date
          description: Date de mise en production
        language:
          type: string
          maxLength: 50
          description: Langage de programmation principal
        editeur:
          type: string
          maxLength: 50
          description: Éditeur ou développeur de l'application
        description:
          type: string
          description: Description détaillée de l'application
      required:
        - nom_application
        - statut
        - type
        - hebergement
        - architecture_logicielle
      example:
        id_application: 550e8400-e29b-41d4-a716-446655440001
        nom_application: CRM Finance
        statut: Run
        type: Build
        hebergement: Cloud
        architecture_logicielle: MVC
        date_mise_en_prod: '2024-01-15'
        language: JavaScript
        editeur: InternalDev
        description: >-
          Application de gestion de la relation client pour le département
          finance
        date_creation: '2024-01-10T08:30:00Z'
        date_modification: '2024-01-15T14:45:00Z'
...
    ThematiqueScore:
      type: object
      properties:
        id_score:
          type: string
          format: uuid
          description: Identifiant unique du score
        thematique:
          type: string
          maxLength: 50
          description: Nom de la thématique
        score:
          type: number
          format: float
          description: Score pour cette thématique
        nombre_reponses:
          type: integer
          description: Nombre de réponses pour cette thématique
        id_analyse:
          type: string
          format: uuid
          description: Identifiant de l'analyse
      required:
        - id_analyse
        - thematique
        - score
        - nombre_reponses
      example:
        id_score: 550e8400-e29b-41d4-a716-446655440007
        id_analyse: 550e8400-e29b-41d4-a716-446655440006
        thematique: Sécurité
        score: 80.5
        nombre_reponses: 15
...
    Formulaire:
      type: object
      properties:
        id_formulaire:
          type: string
          format: uuid
          description: Identifiant unique du formulaire
        id_application:
          type: string
          format: uuid
          description: Identifiant de l'application concernée
        id_questionnaire:
          type: string
          format: uuid
          description: Identifiant du questionnaire utilisé
        id_acteur:
          type: string
          format: uuid
          description: Identifiant de l'acteur qui a rempli le formulaire
        statut:
          type: string
          enum:
            - Brouillon
            - Soumis
            - Validé
          description: Statut du formulaire
        date_creation:
          type: string
          format: date-time
          description: Date de création du formulaire
        date_modification:
          type: string
          format: date-time
          description: Date de dernière modification du formulaire
      required:
        - id_acteur
        - id_application
        - id_questionnaire
        - date_creation
        - date_modification
      example:
        id_formulaire: 550e8400-e29b-41d4-a716-446655440004
        id_acteur: 550e8400-e29b-41d4-a716-446655440000
        id_application: 550e8400-e29b-41d4-a716-446655440001
        id_questionnaire: 550e8400-e29b-41d4-a716-446655440002
        date_creation: '2025-04-20T14:30:00Z'
        date_modification: '2025-04-21T10:15:00Z'
        statut: Validé
...
    MaturityAnalysis:
      type: object
      properties:
        id_analyse:
          type: string
          format: uuid
          description: Identifiant unique de l'analyse
        id_application:
          type: string
          format: uuid
          description: Identifiant de l'application analysée
        date_analyse:
          type: string
          format: date-time
          description: Date de l'analyse
        score_global:
          type: number
          format: float
          description: Score global de maturité
      required:
        - id_application
      example:
        id_analyse: 550e8400-e29b-41d4-a716-446655440006
        id_application: 550e8400-e29b-41d4-a716-446655440001
        date_analyse: '2025-04-22T09:00:00Z'
        score_global: 375.5
    HistoriqueScore:
      type: object
      properties:
        id_historique:
          type: string
          format: uuid
          description: Identifiant unique de l'entrée historique
        id_application:
          type: string
          format: uuid
          description: Identifiant de l'application
        thematique:
          type: string
          maxLength: 50
          description: Nom de la thématique
        score:
          type: number
          format: float
          description: Score pour cette thématique
        date_mesure:
          type: string
          format: date
          description: Date de la mesure
      required:
        - id_application
        - thematique
        - score
        - date_mesure
      example:
        id_historique: 550e8400-e29b-41d4-a716-446655440008
        id_application: 550e8400-e29b-41d4-a716-446655440001
        thematique: Sécurité
        score: 80.5
        date_mesure: '2025-04-22'
...
paths:
...
  /api/applications:
    get:
      summary: Récupère la liste des applications
      tags:
        - Applications
      responses:
        '200':
          description: Liste des applications récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
        '500':
          description: Erreur serveur
    post:
      summary: Crée une nouvelle application
      tags:
        - Applications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
      responses:
        '201':
          description: Application créée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: Données invalides
        '500':
          description: Erreur serveur
  /api/applications/{idApplication}:
    get:
      summary: Récupère une application par son ID
      tags:
        - Applications
      parameters:
        - in: path
          name: idApplication
          required: true
          schema:
            type: string
            format: uuid
          description: Identifiant de l'application
      responses:
        '200':
          description: Application récupérée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '404':
          description: Application non trouvée
        '500':
          description: Erreur serveur
    put:
      summary: Met à jour une application
      tags:
        - Applications
      parameters:
        - in: path
          name: idApplication
          required: true
          schema:
            type: string
            format: uuid
          description: Identifiant de l'application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Application'
      responses:
        '200':
          description: Application mise à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: Données invalides
        '404':
          description: Application non trouvée
        '500':
          description: Erreur serveur
    delete:
      summary: Supprime une application
      tags:
        - Applications
      parameters:
        - in: path
          name: idApplication
          required: true
          schema:
            type: string
            format: uuid
          description: Identifiant de l'application
      responses:
        '200':
          description: Application supprimée avec succès
        '404':
          description: Application non trouvée
        '500':
          description: Erreur serveur
...
  /api/interpretation/application/{idApplication}:
    get:
      summary: Récupère l'analyse complète pour une application
      tags:
        - Analyses
      parameters:
        - in: path
          name: idApplication
          required: true
          schema:
            type: string
            format: uuid
          description: Identifiant de l'application
      responses:
        '200':
          description: Analyse complète récupérée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  idAnalyse:
                    type: string
                    format: uuid
                  idApplication:
                    type: string
                    format: uuid
                  nomApplication:
                    type: string
                  scoreGlobal:
                    type: number
                    format: float
                  interpretation:
                    type: object
                    properties:
                      niveau:
                        type: string
                      description:
                        type: string
                      recommandations:
                        type: string
                      score:
                        type: number
                        format: float
                  thematiques:
                    type: array
                    items:
                      type: object
                      properties:
                        thematique:
                          type: string
                        score:
                          type: number
                          format: float
                  dateAnalyse:
                    type: string
                    format: date-time
        '404':
          description: Application non trouvée
        '500':
          description: Erreur serveur
...
  /api/interpretation/llm/application/{idApplication}:
    get:
      summary: Prépare les données d'analyse pour une interface LLM
      tags:
        - Analyses
      parameters:
        - in: path
          name: idApplication
          required: true
          schema:
            type: string
            format: uuid
          description: Identifiant de l'application
      responses:
        '200':
          description: Données préparées avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  application:
                    $ref: '#/components/schemas/Application'
                  analyse:
                    $ref: '#/components/schemas/MaturityAnalysis'
                  thematiques:
                    type: array
                    items:
                      $ref: '#/components/schemas/ThematiqueScore'
                  historique:
                    type: array
                    items:
                      $ref: '#/components/schemas/HistoriqueScore'
                  formulaires:
                    type: array
                    items:
                      $ref: '#/components/schemas/Formulaire'
        '404':
          description: Application non trouvée
        '500':
          description: Erreur serveur
...